+++ packages/nextjs-mf/src/plugins/container/InvertedContainerPlugin.ts
@@ -1,6 +1,5 @@
 import type { Compiler } from 'webpack';
 import EmbeddedContainerPlugin from './EmbeddedContainerPlugin';
-import { AsyncBoundaryPlugin } from '@module-federation/enhanced';
 
 interface InvertedContainerOptions {
   container?: string;
@@ -20,14 +19,6 @@ class InvertedContainerPlugin {
     new EmbeddedContainerPlugin({
       runtime: this.options.runtime,
       container: this.options.container,
-    }).apply(compiler);
-
-    new AsyncBoundaryPlugin({
-      excludeChunk: (chunk) =>
-        chunk.name === this.options.container ||
-        chunk.name === this.options.container + '_partial',
-      // @ts-ignore
-      eager: (module) => /\.federation/.test(module?.request || ''),
     }).apply(compiler);
   }
 }