+++ packages/nextjs-mf/utils/index.ts
@@ -3,20 +3,11 @@
  * @module @module-federation/utilities/src/utils/pure
  */
 export { extractUrlAndGlobal } from '@module-federation/utilities';
-import { injectScript as injectLegacy } from '@module-federation/utilities';
-//@ts-ignore
-export const injectScript = (args) => {
-  console.error(
-    '@module-federation/utilities injectScript is deprecated, use module-federation/runtime {init,loadRemote}',
-  );
-  return injectLegacy(args);
-};
 
 /**
  * Flushes chunks from the module federation node utilities.
  * @module @module-federation/node/utils
  */
-// @ts-ignore
 export { flushChunks } from '@module-federation/node/utils';
 
 /**
@@ -35,16 +26,16 @@ export type { FlushedChunksProps } from './flushedChunks';
  * If the function is called on the server side, it imports the revalidate function from the module federation node utilities and returns the result of calling that function.
  * @returns {Promise<boolean>} A promise that resolves with a boolean.
  */
-export const revalidate = (
+export const revalidate = function (
   fetchModule: any = undefined,
   force: boolean = false,
-) => {
+): Promise<boolean> {
   if (typeof window !== 'undefined') {
     console.error('revalidate should only be called server-side');
     return Promise.resolve(false);
+  } else {
+    return import('@module-federation/node/utils').then(function (utils) {
+      return utils.revalidate(fetchModule, force);
+    });
   }
-  // @ts-ignore
-  return import('@module-federation/node/utils').then((utils) => {
-    return utils.revalidate(fetchModule, force);
-  });
 };