+++ apps/router-demo/router-remote3-2003/CHANGELOG.md
@@ -1,5 +1,14 @@
 # remote3
 
+## 1.0.21
+
+### Patch Changes
+
+- Updated dependencies [2855583]
+- Updated dependencies [2855583]
+  - @module-federation/enhanced@0.6.1
+  - @module-federation/bridge-vue3@0.6.1
+
 ## 1.0.20
 
 ### Patch Changes